Module google_chromemanagement1::api

Structs

ChromeManagement
Central instance to access all ChromeManagement related resource activities

CustomerMethods
A builder providing access to all methods supported on customer resources. It is not used directly, but through the ChromeManagement hub.

CustomerReportCountChromeVersionCall
Generate report of installed Chrome versions.

CustomerReportCountInstalledAppCall
Generate report of app installations.

CustomerReportFindInstalledAppDeviceCall
Generate report of devices that have a specified app installed.

GoogleChromeManagementV1BrowserVersion
Describes a browser version and its install count.

GoogleChromeManagementV1CountChromeVersionsResponse
Response containing requested browser versions details and counts.

GoogleChromeManagementV1CountInstalledAppsResponse
Response containing details of queried installed apps.

GoogleChromeManagementV1Device
Describes a device reporting Chrome browser information.

GoogleChromeManagementV1FindInstalledAppDevicesResponse
Response containing a list of devices with queried app installed.

GoogleChromeManagementV1InstalledApp
Describes an installed app.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.